requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
autopep8==2.3.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
cssutils==2.11.1
databases==0.9.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.5.0
motor==3.6.0
openai==1.59.6
premailer==3.10.0
pycodestyle==2.12.1
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
yagmail==0.15.293
psycopg2-binary
openai
pinecone-client
python-dotenv
